+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
   * `api::common::CommaSeparatedList` now exists for easy use of
     comma-separated values.
+  * Project members can now be removed via `RemoveProjectMember`.
 
 ## Deprecations